requirements.txt
line-bot-sdk==1.19.0
flask==2.0.1
aiml==0.9.2
gunicorn==20.1.0
xlearn==0.40a1
pandas==1.3.5
beautifulsoup4==4.10.0
urllib3==1.26.8
